src/pyproject.toml
[tool.poetry]
name = "researchhub"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
setuptools = "^65.4.0"
Django = "4.1"
postgres = "^4.0"
djangorestframework = "3.14.0"
django-allauth = "0.48.0"
django-contrib-comments = "1.9.2"
flake8 = "5.0.4"
django-cors-headers = "3.5.0"
boto3 = "1.15.3"
django-storages = "1.10.1"
django-filter = "2.4.0"
sentry-sdk = "^1.4.3"
django-extensions = "3.0.9"
smart-open = "5.2.1"
pre-commit = "2.17.0"
django-ses = "1.0.3"
web3 = "^6.3.0"
django-silk = "5.0.2"
django-dbbackup = "4.0.2"
celery = "4.4.7"
django-celery-beat = "2.0.0"
ipdb = "0.13.9"
habanero = "1.0.0"
celery-redbeat = "1.0.0"
Unidecode = "^1.3.2"
python-slugify = "4.0.1"
jellyfish = "0.8.9"
nltk = "3.6.5"
sqlparse = "0.3.1"
watchdog = "2.1.9"
xmltodict = "0.12.0"
elastic-apm = "6.13.2"
arxiv = "0.5.3"
google-api-python-client = "1.12.1"
mailchimp-marketing = "^3.0.72"
Sift = "5.0.1"
django-ipware = "3.0.1"
geoip2 = "4.4.0"
stripe = "2.62.0"
python-twitter = "3.5"
lxml = "4.6.4"
html5lib = "1.1"
flower = "0.9.7"
Pillow = "8.4.0"
feedparser = "6.0.8"
manubot = "0.5.3"
django-celery-results = "2.0.1"
django-elasticsearch-debug-toolbar = "2.0.0"
django-debug-toolbar = "3.1.1"
PyJWT = "2.3.0"
elasticsearch = "7.13.1"
PyMuPDF = "1.21.0"
pandas = "1.3.4"
newrelic = "^7.2.4"
psycopg2-binary = "2.8.6"
iso8601 = "1.0.2"
dj-rest-auth = "^2.2.2"
black = "^22.3.0"
cloudscraper = "^1.2.60"
djangorestframework-api-key = "2.2"
Markdown = "^3.3.7"
bs4 = "^0.0.1"
responses = "0.21.0"
django-mjml = "1.0"
channels = "3.0.5"
daphne = "3.0.2"
channels-redis = "3.4.1"
elasticsearch-dsl = "7.3.0"
django-elasticsearch-dsl = "7.2.2"
django-elasticsearch-dsl-drf = "0.20.8"
q5-django-inlinecss = "0.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
